+++ application/SAASGPJAX_pipeline.py
@@ -4,7 +4,7 @@ from domain.gp_model import SAASGP, SAASGPJAX
 from application.init_pipeline import init_pipeline, get_data, validate_data
 from adapters.pyro.pyro_model_jax import fit_fully_bayesian_model_nuts
 from domain.env import USE_DUMMY_DATA, MODELDIR, LOGDIR, MEAN_FUNC, KERNEL_TYPE, KERNEL_STRUCTURE, SWS, Y, DATA_SLICE_AMOUNT, POLY_DEGREE
-from domain.metrics import get_metrics, get_BIC
+from domain.scores import get_metrics, get_BIC
 import datetime
 from jax import profiler
