+++ application/deprecated/saasgp_jax_pipeline.py
@@ -9,7 +9,7 @@ from adapters.pyro.pyro_model import fit_fully_bayesian_model_nuts
 from domain.env import USE_DUMMY_DATA, MODELDIR, EXTRAFUNCTIONAL_FEATURES, POLY_DEGREE, MEAN_FUNC, KERNEL_TYPE, KERNEL_STRUCTURE, ARD, RESULTS_DIR
 from domain.feature_model.feature_modeling import inverse_map
 from adapters.gpytorch.sampling import get_initial_points, generate_batch, draw_random_samples, draw_random_x, generate_test_x_from_tensor
-from domain.metrics import get_metrics, gaussian_log_likelihood
+from domain.scores import get_metrics, gaussian_log_likelihood
 import numpy as np
 from scipy.special import logsumexp
 from sklearn.metrics import mean_absolute_percentage_error, r2_score, explained_variance_score
