+++ application/fully_bayesian_gp.py
@@ -4,7 +4,7 @@ from domain.gp_model import SAASGP
 from application.init_pipeline import init_pipeline
 from adapters.pyro.pyro_model import fit_fully_bayesian_model_nuts
 from domain.env import USE_DUMMY_DATA, MODELDIR, MEAN_FUNC, KERNEL_TYPE, KERNEL_STRUCTURE, SWS, Y, DATA_SLICE_AMOUNT, POLY_DEGREE
-from domain.metrics import get_metrics, get_BIC
+from domain.scores import get_metrics, get_BIC
 import datetime
