+++ application/gpytorch_pipeline.py
@@ -4,7 +4,7 @@ from domain.gp_model import MyExactGP, MyApproximateGP
 from gpytorch.mlls import ExactMarginalLogLikelihood, InducingPointKernelAddedLossTerm, VariationalELBO, GammaRobustVariationalELBO
 from application.init_pipeline import init_pipeline, get_data, validate_data
 from domain.env import USE_DUMMY_DATA, MODELDIR, MEAN_FUNC, KERNEL_TYPE, KERNEL_STRUCTURE, DATA_SLICE_AMOUNT, SWS, Y, POLY_DEGREE
-from domain.metrics import get_metrics, gaussian_log_likelihood
+from domain.scores import get_metrics, gaussian_log_likelihood
 import numpy as np
 import datetime
 from time import time
